requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
Django==5.2.7
django-environ==0.12.0
django-jazzmin==3.0.1
idna==3.11
msal==1.34.0
pillow==12.0.0
psycopg2==2.9.11
pycparser==2.23
PyJWT==2.10.1
qrcode==8.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
